+++ apps/prairielearn/src/pages/instructorStudents/instructorStudents.html.tsx
@@ -21,19 +21,17 @@ import { formatDate } from '@prairielearn/formatter';
 import { run } from '@prairielearn/run';
 import {
   CategoricalColumnFilter,
+  NuqsAdapter,
   OverlayTrigger,
   TanstackTableCard,
   TanstackTableEmptyState,
+  parseAsColumnPinningState,
+  parseAsColumnVisibilityStateWithColumns,
+  parseAsSortingState,
 } from '@prairielearn/ui';
 
 import { EnrollmentStatusIcon } from '../../components/EnrollmentStatusIcon.js';
 import { FriendlyDate } from '../../components/FriendlyDate.js';
-import {
-  NuqsAdapter,
-  parseAsColumnPinningState,
-  parseAsColumnVisibilityStateWithColumns,
-  parseAsSortingState,
-} from '../../lib/client/nuqs.js';
 import type { PageContext, PageContextWithAuthzData } from '../../lib/client/page-context.js';
 import { type StaffEnrollment, StaffEnrollmentSchema } from '../../lib/client/safe-db-types.js';
 import { QueryClientProviderDebug } from '../../lib/client/tanstackQuery.js';
@@ -429,6 +427,7 @@ function StudentsCard({
                 : null,
             };
           },
+          hasSelection: false,
         }}
         headerButtons={
           <>
@@ -448,8 +447,6 @@ function StudentsCard({
           </>
         }
         globalFilter={{
-          value: globalFilter,
-          setValue: setGlobalFilter,
           placeholder: 'Search by UID, name, email...',
         }}
         tableOptions={{
@@ -515,7 +512,7 @@ export const InstructorStudents = ({
   search: string;
   isDevMode: boolean;
 } & StudentsCardProps) => {
-  const queryClient = new QueryClient();
+  const [queryClient] = useState(() => new QueryClient());
 
   return (
     <NuqsAdapter search={search}>